+++ piou/__init__.py
@@ -1,3 +1,4 @@
 from .cli import Cli, CommandGroup
 from .utils import Option, Derived
+from .command import CommandMeta
 from .exceptions import CommandNotFoundError